scrapers/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
supabase==2.4.0
openai==1.12.0
anthropic==0.18.1
python-dotenv==1.0.1
pandas==2.2.0